Refactored dashboard.jsx to separate concerns and responsibilities
