+++ frontend/src/functionality/Dashboard/DashboardError.jsx
@@ -1,0 +1,12 @@
+/**
+ * Displays a dashboard-level error message.
+ *
+ * @param {Object} props
+ * @param {string} props.error - The error message to display.
+ * @returns {JSX.Element}
+ */
+function DashboardError({ error }) {
+    return <p>{error}</p>;
+}
+
+export default DashboardError;
