+++ frontend/src/functionality/Dashboard/stylesheets/TaskSearchBar.css
